start scoring spares
Started the scoring of spare frames. The e2e test for it does not pass
yet, so it is marked with 'skip'.

The validation of allowable input now
allows a slash to be entered.

The entry of a slash does not disrupt progression through the game.

The 11th frame is needed for extra throws when the final throw of the
10th frame is a spare or a strike. It should really be a different
design than the others, but we don't have a test that requires that
yet, so we're just dropping an extra Frame in for the moment.

Scoring breaks completely once a spare has been entered.