reduce complication of scoring logic
The score resolver used a lot of helper functions to grab information
about the previous and next frames. Instead of iterating over frames,
it now iterates over a view of frames, including the previous and next.
This makes it substantially easier to follow the logic.

Another complexity was the need to constantly cast the incoming values
to numbers. That's unavoidable, but now that casting is centralized in
an add function with a simple interface. The internals of that function
are gross, but the call sites are much easier to read now.