prevent second throw strikes
The code already prevented this because of the nature of number casting
and NaN condition checks. There's a new unit test to verify that neither
'x' or 'X' are valid in the second input. Since this logic will need to
change later to allow second throw strikes in some cases, it's not a
priority to make this validation explicit at this time.